Removed token conversion inside contract

+++ dao-protocol/contracts/libraries/LibHackFund.sol
@@ -45,7 +45,7 @@ library LibHackFund {
         uint8 firstRunnerUpPercentage;
         uint8 secondRunnerUpPercentage;
         uint funding;
-        uint16 minScoreTokenRequired;
+        uint minScoreTokenRequired;
         State state;
     }
